+++ IwEngine/src/graphics/VoxelLight.cpp
@@ -72,12 +72,12 @@ namespace Graphics {
 			source->SetAsTexture(m_voxelTexture->Handle(), 0);
 
 			glm::vec3 dim(m_voxelTexture->Width(), m_voxelTexture->Height(), m_voxelTexture->Depth());
-			unsigned levels = log2(glm::max(glm::max(dim.x, dim.y), dim.z)) + 1;
+			unsigned levels = (unsigned)log2(glm::max(glm::max(dim.x, dim.y), dim.z)) + 1;
 
 			int xSize, ySize, zSize;
 			m_mipmapGenerationShader->Handle()->GetComputeWorkGroupSize(xSize, ySize, zSize);
 
-			for (int mip = 1; mip < levels; mip++) {
+			for (unsigned mip = 1; mip < levels; mip++) {
 				level ->SetAsInt(mip - 1);
 				target->SetAsImage(m_voxelTexture->Handle(), 1, mip);
